requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
python-dotenv==1.0.1
requests==2.32.3
setuptools==72.2.0
urllib3==2.2.2
wheel==0.44.0